Add tests for LighthouseChart
