+++ Lighthouse.Frontend/setupTests.ts
@@ -1,1 +1,4 @@
 import '@testing-library/jest-dom';
+import { registerLicense } from '@syncfusion/ej2-base';
+
+registerLicense('ORg4AjUWIQA/Gnt2UlhhQlVMfV5AQmBIYVp/TGpJfl96cVxMZVVBJAtUQF1hTX9Td0RjUHxcc3xTQ2Bd');
