+++ Lighthouse.Frontend/src/pages/Projects/Detail/SampleFrequencySelector.test.tsx
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi, afterEach } from 'vitest';
-import { fireEvent, render, screen, waitFor, within } from '@testing-library/react';
+import { fireEvent, render, screen, within } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import SampleFrequencySelector from './SampleFrequencySelector';
